% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_pg.R
\name{connect_to_pg}
\alias{connect_to_pg}
\title{Create a pool connection to a PostgreSQL database either using a config file or environment variables
If both exist, config file is given priority}
\usage{
connect_to_pg(
  config_file = here::here("config.yml"),
  config_value = "default",
  envars_prefix = "DEFAULT_DATABASE_",
  ...
)
}
\arguments{
\item{config_file}{A yml config file with database credentials. Default is \code{config.yml} in root of project}

\item{config_value}{The set of config vars to use from the config file}

\item{envars_prefix}{The prefix of the environment variables to use to connect}

\item{...}{Arguments passed to config::get}
}
\value{
A database connection object
}
\description{
Create a pool connection to a PostgreSQL database either using a config file or environment variables
If both exist, config file is given priority
}
