% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lazy_conn.R
\name{create_lazy_conn}
\alias{create_lazy_conn}
\title{Create a lazy connection to a table in a schema}
\usage{
create_lazy_conn(pool_conn, schemaname = "public", tablename)
}
\arguments{
\item{pool_conn}{A pool connection to a PostgreSQL database}

\item{schemaname}{The schema from which to get the tables. Default is public}

\item{tablename}{The name of the table to create the connection to}
}
\value{
A pool connection object
}
\description{
Create a lazy connection to a table in a schema
}
