% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_merge_into.R
\name{db_merge_into}
\alias{db_merge_into}
\title{Merge data into SQL database}
\usage{
db_merge_into(db, table, values, by, id_colname, backend = "db_insert_into",
  ...)
}
\arguments{
\item{db}{dplyr \code{src_*} object}

\item{table}{Name of table, as character}

\item{values}{\code{data.frame}}

\item{by}{Character vector of column names to use for the merge}

\item{id_colname}{Column name for the database ID}

\item{backend}{Which backend to use for copy operation. Default is "db_insert_into".}

\item{...}{Additional arguments to \code{backend_*} function}
}
\description{
Merge data into SQL database
}
\examples{
data(iris)
tablename <- 'projects'
input <- data.table(
  code = 'test',
  description = 'example')
key <- 'code'
dbname <- 'leaf_spectra'
merge_with_sql(input, tablename, key = 'code')
}
\author{
Alexey Shiklomanov <ashiklom@bu.edu>
}

