% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_where.R
\name{db_update_where}
\alias{db_update_where}
\title{Update SQL table based on criteria}
\usage{
db_update_where(db, table, criteria, values)
}
\arguments{
\item{db}{dplyr \code{src_*} object}

\item{table}{Name of table, as character}

\item{criteria}{(list of lists) See details.}

\item{values}{Named list of values. See details}
}
\description{
Function to facilitate updating SQL database.
}
\details{
Each item in criteria should be a length 3 list containing:
 * The column name, as a character
 * The operator (e.g. '=', '<'), as a string
 * The value, as the appropriate type
}
