% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2sql_write.R
\name{dt2sql_write}
\alias{dt2sql_write}
\title{Write data.table to SQL}
\usage{
dt2sql_write(db, tablename, input, id_colname = "id", add_id = TRUE, ...)
}
\description{
Write data.table to SQL
}
\examples{
library(specprocess)
tablename <- 'projects'
input <- data.table(
   id = 1e+08,
   code = 'test3',
   description = 'example')
id_colname <- 'id'
add_id <- TRUE
db <- src_postgres('leaf_spectra')
con <- db$con
dt2sql_write(db, tablename, input, key = 'code', add_id = FALSE)
}

