% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_filename.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate a file name for a report}
\usage{
generate_filename(
  reports_dir = "~/.reports/",
  report_name,
  date,
  module,
  extension = ".csv",
  device_name,
  separator = "_"
)
}
\arguments{
\item{reports_dir}{Directory to prefix the report name with. Default is ~/.reports/}

\item{report_name}{Name of the report}

\item{date}{Date of the report}

\item{module}{Module of the report}

\item{extension}{File extension. Default is ".csv"}

\item{device_name}{Name of the device the report comes from}

\item{separator}{Separator between the inputs. Default is "_"}
}
\value{
A string containing the generated file name in the format:
  reports_dir sep
  device_name sep
  report_name sep
  module sep
  date
  extension

  e.g ~/.reports/monthend_KRP_numeracy_06-20.csv
}
\description{
Generate a file name for a report
}
\examples{
generate_filename(
  report_name="monthend",
  date="06-20",
  module="numeracy",
  device_name = "KRP")

generate_filename(
  report_name="alldata",
  date="06-20",
  module="literacy",
  device_name = "DRY")
}
