% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tables_data.R
\name{get_all_tables_data}
\alias{get_all_tables_data}
\title{Get data from all tables in a schema. Each as a separate \code{data.frame}}
\usage{
get_all_tables_data(
  pool_conn,
  schemaname = "public",
  prefix = "",
  env = globalenv()
)
}
\arguments{
\item{pool_conn}{A pool connection to a PostgreSQL database}

\item{schemaname}{The schema from which to get the tables. Default is public}

\item{prefix}{String to prefix the variables with}

\item{env}{The environment in which to create the variables. Default is global environment}
}
\value{
A \code{data.frame} for each of the tables in the specified schema
}
\description{
Get data from all tables in a schema. Each as a separate \code{data.frame}
}
