% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_name.R
\name{get_first_name}
\alias{get_first_name}
\title{Get the first name from a full name}
\usage{
get_first_name(full_name)
}
\arguments{
\item{full_name}{The full name from which to extract the first name}
}
\value{
A string containing the first name
}
\description{
Get the first name from a full name
}
\examples{
get_first_name("John Nsombo")
get_first_name("Jorgen Von Straggle")
}
