% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_name.R
\name{get_last_name}
\alias{get_last_name}
\title{Get the last name from a full name}
\usage{
get_last_name(full_name)
}
\arguments{
\item{full_name}{The full name from which to extract the last name}
}
\value{
A string containing the last name. All names after the first name will be taken as the last name
}
\description{
Get the last name from a full name
}
\examples{
get_last_name("John Nsombo")
get_last_name("Jorgen Von Straggle")
}
