% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{get_table_data}
\alias{get_table_data}
\title{Get data from a database table as a \code{data.frame}}
\usage{
get_table_data(pool_conn, schemaname = "public", tablename)
}
\arguments{
\item{pool_conn}{A pool connection to a PostgreSQL database}

\item{schemaname}{The schema from which to get the tables. Default is public}

\item{tablename}{The name of the table to create the connection to}
}
\value{
A \code{data.frame} containing the data in the specified table
}
\description{
Get data from a database table as a \code{data.frame}
}
