% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Get the mode from a vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{A vector}
}
\value{
A vector containing the most frequent value in a vector.
If there is a tie, the top values will be returned.
R currently has no built-in function to get mode
}
\description{
Get the mode from a vector
}
\examples{
hps <- mtcars$hp
getmode(hps)

}
