% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_conn_all_tables.R
\name{lazy_conn_all_tables}
\alias{lazy_conn_all_tables}
\title{Create lazy connections to all tables in a schema as separate variables}
\usage{
lazy_conn_all_tables(
  pool_conn,
  schemaname = "public",
  prefix = "",
  env = globalenv()
)
}
\arguments{
\item{pool_conn}{A pool connection to a PostgreSQL database}

\item{schemaname}{The schema from which to get the tables. Default is public}

\item{prefix}{String to prefix the variables with}

\item{env}{The environment in which to create the variables. Default is global environment}
}
\value{
Variables for lazy connections to each of the tables in the environment specified
}
\description{
Create lazy connections to all tables in a schema as separate variables
}
