% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_uuid.R
\name{str_to_uuid}
\alias{str_to_uuid}
\title{Convert a string to a uuid by inserting hyphens in the right places}
\usage{
str_to_uuid(input_string)
}
\arguments{
\item{input_string}{A 32-character string to convert to a uuid}
}
\value{
A string formatted as a uuid
}
\description{
Convert a string to a uuid by inserting hyphens in the right places
}
\examples{
str_to_uuid("f3f6bf4b9c424b6fbd90bece1418a415")
str_to_uuid("3a126f041a9e4127978bb6544a099d0d")

}
